refactor(nmide): Moved concepts to specific files
