+++ nmide/lib/Event.hs
@@ -1,0 +1,9 @@
+module Event where
+
+import Value (Value)
+
+data Event = Event
+  { moduleName :: String
+  , eventName :: String
+  , arguments :: Maybe Value
+  }
