+++ nmide/lib/Mod.hs
@@ -1,0 +1,17 @@
+module Mod where
+
+data Mod a
+  -- Adds a Node
+  = Add
+  { node :: a
+  , pred :: a -> Bool
+  }
+  -- Removes a Node
+  | Rem
+  { pred :: a -> Bool
+  }
+  -- Modifies a Node
+  | Mod
+  { pred :: a -> Bool
+  , mod :: a -> a
+  }
